+++ src/Plugin.cpp
@@ -1,116 +1,125 @@
+#include "Geode/utils/file.hpp"
 #include <internal/SerpentLua.hpp>
 
 using namespace geode::prelude;
 using namespace SerpentLua;
 
 void Plugin::terminate() {
-	if (native) {
-		if (hDll.has_value()) {
-			FreeLibrary(hDll.value());
-		}
-	}
-	auto annoyingNoDiscardWarningFix = internal::RuntimeManager::get()->removeLoadedPlugin(this->metadata->id);
-	if (annoyingNoDiscardWarningFix.isErr()) {
-		log::warn("{}", annoyingNoDiscardWarningFix.err().value());
-		return;
-	}
+	(void)internal::RuntimeManager::get()->removeLoadedPlugin(this->metadata->id); // this only returns an error when the plugin isnt registered at all
+	if (this->state) lua_close(this->state);
+	
 	delete this;
 }
 
-std::function<void(lua_State*)> Plugin::getEntry() {
+std::function<bool()> Plugin::getEntry() {
 	return entry;
+}
+
+std::function<bool(lua_State*)> Plugin::getOnScriptLoaded() {
+	return onScriptLoaded;
+}
+
+lua_State* Plugin::getLuaState() {
+	return state;
 }
 
 void Plugin::setPlugin() {
 	return internal::RuntimeManager::get()->setPlugin(this);
 }
 
-void Plugin::SerpentLuaAPIImpl::log(__metadata m, const char* msg, const char* type) {
-	if (std::strcmp(type, "info") == 0) log::info("[PLUGIN] [{}]: {}", m.name, msg);
-	else if (strcmp(type, "warn") == 0) log::warn("[PLUGIN] [{}]: {}", m.name, msg);
-	else if (strcmp(type, "error") == 0) log::error("[PLUGIN] [{}]: {}", m.name, msg);
-	else if (strcmp(type, "debug") == 0) log::debug("[PLUGIN] [{}]: {}", m.name, msg);
-	else log::trace("[PLUGIN] [{}]: {}", m.name, msg); // always default to trace if type isnt those things
-}
-
-Plugin::__metadata Plugin::SerpentLuaAPIImpl::get_script(lua_State* L) {
-	auto script = SerpentLua::internal::RuntimeManager::get()->getScriptByState(L);
-
-	Plugin::__metadata ret;
-	ret.name = script->name.c_str();
-	ret.developer = script->developer.c_str();
-	ret.id = script->id.c_str();
-	ret.version = script->version.c_str();
-	ret.serpentVersion = script->serpentVersion.c_str();
-
-	std::vector<const char*> plugins;
-	plugins.reserve(script->plugins.size());
-	for (const auto& [butt, plug] : script->plugins) {
-		plugins.push_back(butt.c_str());
-	}
-	ret.pluginsSize = plugins.size();
-	ret.plugins = plugins.data();
-
-	return ret;
-}
-
 geode::Result<Plugin*, std::string> Plugin::createNative(const std::filesystem::path& path) {
 	log::info("Loading Native Plugin {}: initialized", path.filename());
 	if (!Mod::get()->getSavedValue<bool>(fmt::format("safe-{}", path.stem())) && !Mod::get()->getSettingValue<bool>("dev-mode")) return Err("Native Plugin {} was imported manually.\nThis plugin will not load unless it's imported through the plugin importer in-game.", path.stem());
+	auto configDir = Mod::get()->getConfigDir();
+	auto unzipped = configDir / "unzipped" / path.stem();
+	bool skipUnzip = false;
 
-	auto temphDll = LoadLibraryExW(path.c_str(), NULL, LOAD_LIBRARY_AS_DATAFILE);
-	if (!temphDll) {
-		return Err("Plugin {}: Failed to load SLP.", path.filename());
+	if (std::filesystem::exists(unzipped)) {
+		auto originalHashRes = utils::file::readString(unzipped / ".slp_sha256");
+		if (originalHashRes.isOk()) {
+			auto originalHash = originalHashRes.unwrap();
+
+			auto binaryRes = utils::file::readBinary(path);
+			if (binaryRes.isErr()) {
+				return Err("Plugin {}: Unable to read SLP: {}", path.filename(), binaryRes.unwrapErr()); // if this fails then i dont think unzipping would work either
+			}
+			auto binary = binaryRes.unwrap();
+
+			auto hash = internal::utility::sha256(binary);
+
+			skipUnzip = hash == originalHash;
+		} else {
+			skipUnzip = false;
+			log::warn("Plugin {}: Unable to read .slp_sha256: Will unzip anyway.", path.filename());
+		}
 	}
+
+
+	log::debug("Plugin {}: Unpacking SLP...", path.filename());
+
+	if (!skipUnzip) {
+		log::debug("Plugin {}: Hash mismatch, unzipping.", path.filename());
+		std::filesystem::remove_all(unzipped);
+		auto unzip = utils::file::Unzip::intoDir(path, unzipped);
+		if (unzip.isErr()) return Err("Plugin: Unable to unpack SLP: {}", unzip.unwrapErr());
+
+		auto binary = utils::file::readBinary(path).unwrap(); // if it could unzip then this shouldnt error (hopefully)
+		auto hash = internal::utility::sha256(binary);
+
+		auto write = utils::file::writeString(unzipped / ".slp_sha256", hash);
+		if (write.isErr()) {
+			log::warn("Plugin {}: Couldn't write .slp_sha256: {}. Will continue loading regardless.", path.filename(), write.unwrapErr());
+		}
+	} else {
+		log::debug("Plugin {}: Same hash, skipping unzip.", path.filename());
+	}
+
+	auto initRes = utils::file::readString(unzipped / "init.lua");
+	if (initRes.isErr()) return Err("Plugin {}: Unable to read init script: {}", path.filename(), initRes.unwrapErr());
 
 	log::debug("Plugin {}: SLP loaded, gathering metadata...", path.filename());
 
-	auto metaRes = PluginMetadata::createFromSLP(path, temphDll);
-	if (metaRes.isErr()) return Err("Gathering Metadata: {}", metaRes.unwrapErr());
+	auto metaRes = PluginMetadata::createFromScript(unzipped);
+	if (metaRes.isErr()) return Err("Plugin {}: Gathering Metadata: {}", path.filename(), metaRes.unwrapErr());
 	auto metadata = metaRes.unwrap();
-
-	FreeLibrary(temphDll);
 
 	log::debug("Plugin {}: Metadata gathered, executing...", path.filename());
 
+	auto state = luaL_newstate();
+	luaL_openlibs(state); // plugins were never sandboxed anyway
 
-	auto hDll = LoadLibraryExW(path.c_str(), NULL, LOAD_LIBRARY_SEARCH_USER_DIRS);
-
-	if (!hDll) return Err("Plugin {}: Failed to load with error {}", path.filename(), GetLastError());
-
-	// now we... DONT handle METADATA? it has already retrieved metadata. now we.. handle API FUNCTIONS...!
-	log::debug("Plugin {}: SLP executed.", path.filename());
-
-	auto INITFUCKINGAPI = reinterpret_cast<void(*)(Plugin::SerpentLuaAPI)>(GetProcAddress(hDll, "initNativeAPI"));
-	if (!INITFUCKINGAPI) {
-		FreeLibrary(hDll);
-		// return Err("THE FUCKING FUNCTION WASNT FOUND")
-		return Err("Plugin {}: initNativeAPI function was not found.", path.filename());
+	if (luaL_loadfile(state, utils::string::pathToString((unzipped / "init.lua")).c_str()) != LUA_OK) {
+		auto err = Err("Plugin {}: Compile error: {}", path.filename(), lua_tostring(state, -1));
+		lua_close(state);
+		return err;
 	}
 
-	Plugin::__metadata __md;
-	__md.name = metadata->name.c_str();
-	__md.developer = metadata->developer.c_str();
-	__md.id = metadata->id.c_str();
-	__md.version = metadata->version.c_str();
-	__md.serpentVersion = metadata->serpentVersion.c_str();
-	__md.plugins = {}; // just initializing it to empty to be safe
-
-	Plugin::SerpentLuaAPI API;
-	API.handle = hDll;
-	API.metadata = __md;
-	API.log = &Plugin::SerpentLuaAPIImpl::log;
-	API.get_script = &Plugin::SerpentLuaAPIImpl::get_script;
-
-	INITFUCKINGAPI(API);
-
-	auto rawEntry = reinterpret_cast<void(*)(lua_State*)>(GetProcAddress(hDll, "entry"));
-	if (!rawEntry) {
-		FreeLibrary(hDll);
-		return Err("Plugin {}: Entry function was not found within the SLP.", path.filename());
+	if (lua_pcall(state, 0, 1, 0) != LUA_OK) {
+		auto err = Err("Plugin {}: Runtime error: {}", path.filename(), lua_tostring(state, -1));
+		lua_close(state);
+		return err;
 	}
 
-	auto plugin = Plugin::create(metadata, rawEntry);
+	lua_getfield(state, -1, "entry");
+	int entryRef = luaL_ref(state, LUA_REGISTRYINDEX);
+
+	std::function<bool()> entry = [state, entryRef]() {
+		lua_rawgeti(state, LUA_REGISTRYINDEX, entryRef);
+
+		return lua_pcall(state, 0, 0, 0) == LUA_OK;
+	};
+
+	lua_getfield(state, -1, "onScriptLoaded");
+	int oslRef = luaL_ref(state, LUA_REGISTRYINDEX);
+
+	std::function<bool(lua_State*)> onScriptLoaded = [state, oslRef](lua_State* scriptState) {
+		lua_rawgeti(state, LUA_REGISTRYINDEX, oslRef);
+		lua_pushlightuserdata(state, scriptState);
+
+		return lua_pcall(state, 1, 0, 0) == LUA_OK;
+	};
+
+	auto plugin = Plugin::create(metadata, state, entry, onScriptLoaded);
 	if (plugin.isErr()) {
 		return Err("Plugin {}: {}", path.filename(), plugin.err());
 	};
@@ -120,17 +129,26 @@ geode::Result<Plugin*, std::string> Plugin::createNative(const std::filesystem::
 	metadata->native = true;
 	metadata->loaded = true;
 
-	unwrapped->hDll = hDll;
 	return Ok(unwrapped);
 }
 
-Result<Plugin*, std::string> Plugin::create(PluginMetadata* metadata, std::function<void(lua_State*)> entry) {
-	log::info("Plugin {} creation: initialized.", metadata->id);
+Result<Plugin*, std::string> Plugin::create(PluginMetadata* metadata, lua_State* state, std::function<bool()> entry, std::function<bool(lua_State*)> onScriptLoaded) {
+	log::info("Plugin {}: initialized.", metadata->id);
 	auto ret = new (std::nothrow) Plugin();
-	if (!ret) return Err("Plugin creation: Not enough memory to create plugin.");
+	if (!ret) return Err("Plugin {}: Not enough memory to create plugin.", metadata->id);
 	ret->loadCount = 0;
 
+	ret->state = state;
 	ret->metadata = metadata;
 	ret->entry = entry;
+	ret->onScriptLoaded = onScriptLoaded;
+
+	if (!ret->getEntry()()) {
+		auto err = Err("Plugin {}: Entry function errored: {}", metadata->id, lua_tostring(ret->state, -1));
+		lua_pop(ret->state, -1);
+		ret->terminate();
+		return err;
+	}
+
 	return Ok(ret);
 }